requirements.txt
Flask
flask-cors
bcrypt
psycopg2-binary
